+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # main
 
+* Added cop `Migration/ChangeTableReferences` ([#47](https://github.com/petalmd/rubocop-petal/pull/47))
 * Added cop `Migration/AlwaysBulkChangeTable` ([#46](https://github.com/petalmd/rubocop-petal/pull/46))
 
 # v1.0.0 (2023-05-15)